refactor: seperate view-all, add, edit and delete pages
