+++ src/pages/albums/add.js
@@ -1,0 +1,7 @@
+const { setTitle } = require('../../components/index.js');
+
+function buildAddAllPage() {
+  setTitle('Add a new album');
+}
+
+module.exports = buildAddAllPage;
