+++ src/pages/albums/delete.js
@@ -1,0 +1,7 @@
+const { setTitle } = require('../../components/index.js');
+
+function buildDeletePage() {
+  setTitle('Delete a album');
+}
+
+module.exports = buildDeletePage;
