+++ src/pages/albums/edit.js
@@ -1,0 +1,7 @@
+const { setTitle } = require('../../components/index.js');
+
+function buildEditAllPage() {
+  setTitle('Edit a album');
+}
+
+module.exports = buildEditAllPage;
