+++ src/pages/albums/viewAll.js
@@ -1,0 +1,7 @@
+const { setTitle } = require('../../components/index.js');
+
+function buildAlbumsAllPage() {
+  setTitle('Overview of all albums');
+}
+
+module.exports = buildAlbumsAllPage;
