+++ src/pages/cds/add.js
@@ -1,0 +1,8 @@
+const { setTitle } = require('../../components/index.js');
+
+
+function buildCDsAddPage() {
+  setTitle('Add a new CD to your collection');
+}
+
+module.exports = buildCDsAddPage;
