+++ src/pages/cds/delete.js
@@ -1,0 +1,7 @@
+const { setTitle } = require('../../components/index.js');
+
+function buildCDsDeletePage() {
+  setTitle('Remove a CD from your collection');
+}
+
+module.exports = buildCDsDeletePage;
