+++ src/pages/cds/edit.js
@@ -1,0 +1,7 @@
+const { setTitle } = require('../../components/index.js');
+
+function buildCDsEditPage() {
+  setTitle('Edit a CD');
+}
+
+module.exports = buildCDsEditPage;
