+++ src/pages/cds/viewAll.js
@@ -1,0 +1,7 @@
+const { setTitle } = require('../../components/index.js');
+
+function buildCDsAllPage() {
+  setTitle('Overview of all CDs');
+}
+
+module.exports = buildCDsAllPage;
