+++ src/pages/tracks/add.js
@@ -1,0 +1,7 @@
+const { setTitle } = require('../../components/index.js');
+
+function buildTracksAddPage() {
+  setTitle('Add a new track to your collection');
+}
+
+module.exports = buildTracksAddPage;
