+++ src/pages/tracks/delete.js
@@ -1,0 +1,7 @@
+const { setTitle } = require('../../components/index.js');
+
+function buildTracksDeletePage() {
+  setTitle('Remove a track from your collection');
+}
+
+module.exports = buildTracksDeletePage;
