+++ src/pages/tracks/edit.js
@@ -1,0 +1,7 @@
+const { setTitle } = require('../../components/index.js');
+
+function buildTracksEditPage() {
+  setTitle('Edit a track');
+}
+
+module.exports = buildTracksEditPage;
