+++ src/pages/tracks/viewAll.js
@@ -1,0 +1,7 @@
+const { setTitle } = require('../../components/index.js');
+
+function buildTracksAllPage() {
+  setTitle('Overview of all tracks');
+}
+
+module.exports = buildTracksAllPage;
